feat: remove splashv2 and add progressbar toggle
