+++ .eslintrc.json
@@ -64,6 +64,11 @@
 
         "path-alias/no-relative": "error",
 
-        "prettier/prettier": "error"
+        "prettier/prettier": [
+            "error",
+            {
+                "endOfLine": "auto"
+            }
+        ]
     }
 }
