+++ .prettierrc.yaml
@@ -5,4 +5,4 @@ trailingComma: none
 bracketSpacing: true
 arrowParens: avoid
 useTabs: false
-endOfLine: lf
+endOfLine: auto
