+++ src/shared/settings.d.ts
@@ -34,7 +34,7 @@ export interface Settings {
     splashColor?: string;
     splashAnimationPath?: string;
     splashBackground?: string;
-    splashDetailed?: boolean;
+    splashProgress?: boolean;
 
     spellCheckLanguages?: string[];
