+++ static/views/splash.html
@@ -11,16 +11,45 @@
             box-sizing: border-box;
             height: 100%;
             display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
             border-radius: 8px;
             border: 1px solid var(--fg-semi-trans);
         }
 
-        .loading {
-            display: flex;
-            width: 300px;
-            flex-direction: column;
-            justify-content: center;
-            align-items: center;
+        #progress-section {
+            width: 100%;
+        }
+        
+        .progressbar {
+            position: relative;
+            height: 20px;
+            width: auto;
+            box-sizing: border-box;
+            margin-left: 16px;
+            margin-right: 16px;
+            background-color: color-mix(in srgb, var(--bg), black 20%);
+            border: 1px solid color-mix(in srgb, var(--fg-semi-trans), white 10%);
+            border-radius: 4px;
+        }
+        .progressbar #progress {
+            height: 100%;
+            width: 0%;
+            max-width: 100%;
+            background-color: color-mix(in srgb, var(--fg-semi-trans), black 10%);
+            border-radius: 4px;
+            transition: width 0.5s ease-in;
+        }
+
+        .progressbar #progress-percentage {
+            position: absolute;
+            width: 100%;
+            height: 100%;
+            padding-top: 2px;
+            text-align: center;
+            font-size: 13px;
+            font-weight: bold;
         }
 
         p {
@@ -35,11 +64,17 @@
 
 <body>
     <div class="wrapper">
-        <div class="loading">
-            <!-- the data url is here to ensure there isn't an empty frame before the image is loaded -->
-            <img id="animation" src="data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw=="
-                draggable="false" alt="animation" role="presentation" />
-            <p>Loading Equibop...</p>
+        <!-- the data url is here to ensure there isn't an empty frame before the image is loaded -->
+        <img id="animation" src="data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw=="
+        draggable="false" alt="animation" role="presentation" />
+        
+        <p>Loading Equibop...</p>
+
+        <div id="progress-section">
+            <div class="progressbar">
+                <span id="progress-percentage">99%</span>
+                <div id="progress"></div>
+            </div>
         </div>
     </div>
 </body>
